refactor: make Experiments/ the single source of truth for lab code
Consolidate duplicated lab scripts and fix broken GitHub Pages links.

- Remove duplicate .py copies from the module dirs (M1/M2/M3/M5); the
  module dirs now hold only README + images. Where copies differed, the
  Experiments/ version was kept (it is the run directory: it has the shared
  deps like htk_featio.py and the data dirs lists/ and am/).
- Fix misnamed file: Experiments/M3_Train_AM_cntk.py actually contained
  PyTorch -> renamed to M3_Train_AM_torch.py. The genuine CNTK reference
  (only real copy, from M3_Acoustic_Modeling/) is moved into Experiments/
  under the correct M3_Train_AM_cntk.py name.
- Fix 6 broken README links that pointed to ../Experiments/*.py. The
  Experiments/ folder is excluded from the Jekyll site (_config.yml), so
  those resolved to 404s on GitHub Pages. They now point to absolute
  github.com/blob URLs that work both on the site and in the repo.
