+++ Experiments/M3_Train_AM_cntk.py
@@ -1,23 +1,13 @@
-import torch
-import torch.nn as nn
-import torch.optim as optim
-from torch.utils.data import Dataset, DataLoader
+from __future__ import print_function
 import numpy as np
-import os
+import cntk as C
+from cntk.train.training_session import *
+from cntk.logging import *
 import argparse
-from htk_featio import read_htk_user_feat
-from typing import List, Tuple
-import logging
-import time
+import os
 
-# Setup logging
-logging.basicConfig(
-    level=logging.INFO,
-    format='%(asctime)s - %(levelname)s - %(message)s'
-)
-
-# Keep the same data directory structure and globals
-data_dir = "./"
+# model_type = "" # or "blstm"
+data_dir = "../Experiments"
 list_path = os.path.join(data_dir, "lists")
 am_path = os.path.join(data_dir, "am")
 
@@ -34,260 +24,387 @@ globals = {
     "num_classes": 120,
 }
 
-class SpeechDataset(Dataset):
-    def __init__(self, features_file: str, labels_file: str, feature_mean: np.ndarray, 
-                 feature_invstd: np.ndarray, context: Tuple[int, int] = (0,0)):
-        self.features = self.load_features(features_file)
-        self.labels = self.load_labels(labels_file)
-        self.feature_mean = feature_mean
-        self.feature_invstd = feature_invstd
-        self.context = context
-        
-    def load_features(self, file_path: str) -> List[np.ndarray]:
-        features = []
-        with open(file_path, 'r') as f:
-            for line in f:
-                features.append(read_htk_user_feat(line.strip()))
-        return features
-    
-    def load_labels(self, file_path: str) -> List[np.ndarray]:
-        labels = []
-        current_labels = []
-        with open(file_path, 'r') as f:
-            for line in f:
-                line = line.strip()
-                if line == '#!MLF!#':
-                    continue
-                if line.startswith('"'):  # New utterance
-                    if current_labels:
-                        labels.append(np.array(current_labels))
-                    current_labels = []
-                elif line == '.':  # End of utterance
-                    if current_labels:
-                        labels.append(np.array(current_labels))
-                    current_labels = []
-                else:  # Label line
-                    try:
-                        state_id = int(line.split()[0])
-                        current_labels.append(state_id)
-                    except:
-                        continue
-        return labels
-    
-    def __len__(self) -> int:
-        return len(self.features)
-    
-    def __getitem__(self, idx: int) -> Tuple[torch.Tensor, torch.Tensor, int]:
-        feature = self.features[idx]
-        label = self.labels[idx]
-        
-        # Apply mean and variance normalization
-        feature = (feature - self.feature_mean) * self.feature_invstd
-        
-        # Add context frames
-        if self.context != (0,0):
-            left, right = self.context
-            padded_feature = np.pad(feature, ((left, right), (0, 0)), mode='edge')
-            context_features = []
-            for i in range(len(feature)):
-                context_window = padded_feature[i:i+left+right+1]
-                context_features.append(context_window.flatten())
-            feature = np.array(context_features)
-            
-        return torch.FloatTensor(feature), torch.LongTensor(label), len(feature)
 
-class DNNModel(nn.Module):
-    def __init__(self, input_dim: int, hidden_dim: int, num_classes: int, num_layers: int = 4):
-        super(DNNModel, self).__init__()
-        self.layers = nn.ModuleList()
-        
-        # First layer
-        self.layers.append(nn.Linear(input_dim, hidden_dim))
-        self.layers.append(nn.Sigmoid())
-        
-        # Hidden layers
-        for _ in range(num_layers - 1):
-            self.layers.append(nn.Linear(hidden_dim, hidden_dim))
-            self.layers.append(nn.Sigmoid())
-            
-        # Output layer
-        self.layers.append(nn.Linear(hidden_dim, num_classes))
-        
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        for layer in self.layers:
-            x = layer(x)
-        return x
+# frames to left/right of current frame to augment the input with
+# for DNN total input size is 23 frames
+# for RNN, just a single frames is used
 
-class BLSTMModel(nn.Module):
-    def __init__(self, input_dim: int, hidden_dim: int, num_classes: int, num_layers: int = 2, dropout: float = 0.0):
-        super(BLSTMModel, self).__init__()
-        self.lstm = nn.LSTM(
-            input_dim, hidden_dim, num_layers=num_layers, 
-            bidirectional=True, batch_first=True, dropout=dropout
+# helper function to read ascii data files (for mean and inv stddev parameters)
+def load_ascii_vector(file_name, var_name=None):
+    x = np.asarray(np.loadtxt(file_name), dtype=np.float32)
+    par = C.constant(value=x, shape=len(x), name=var_name)
+    return par
+
+
+def create_mb_source(
+    features_file,
+    labels_file,
+    label_mapping_file,
+    feature_dim,
+    num_classes,
+    max_sweeps=C.io.INFINITELY_REPEAT,
+    context=(0, 0),
+    frame_mode=False,
+):
+
+    for file_name in [features_file, labels_file, label_mapping_file]:
+        if not os.path.exists(file_name):
+            raise RuntimeError(
+                "File '{}' does not exist. Please check that datadir argument is set correctly.".format(
+                    file_name
+                )
+            )
+
+    fd = C.io.HTKFeatureDeserializer(
+        C.io.StreamDefs(
+            features=C.io.StreamDef(
+                shape=feature_dim, context=context, scp=features_file
+            )
         )
-        self.fc = nn.Linear(hidden_dim * 2, num_classes)
+    )
 
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        lstm_out, _ = self.lstm(x)
-        return self.fc(lstm_out)
+    ld = C.io.HTKMLFDeserializer(
+        label_mapping_file,
+        C.io.StreamDefs(labels=C.io.StreamDef(shape=num_classes, mlf=labels_file)),
+    )
 
-def train_model(model_type: str, train_loader: DataLoader, val_loader: DataLoader, 
-                device: torch.device, num_epochs: int):
-    # Initialize model
+    return C.io.MinibatchSource([fd, ld], frame_mode=frame_mode, max_sweeps=max_sweeps)
+
+
+def create_network(
+    feature_dim=40,
+    num_classes=256,
+    feature_mean_file=None,
+    feature_inv_stddev_file=None,
+    feature_norm_files=None,
+    label_prior_file=None,
+    context=(0, 0),
+    model_type=None,
+):
+    def MyMeanVarNorm(feature_mean_file, feature_inv_stddev_file):
+        m = C.reshape(
+            load_ascii_vector(feature_mean_file, "feature_mean"), shape=(1, feature_dim)
+        )
+        s = C.reshape(
+            load_ascii_vector(feature_inv_stddev_file, "feature_invstddev"),
+            shape=(1, feature_dim),
+        )
+
+        def _func(operand):
+            return C.reshape(
+                C.element_times(
+                    C.reshape(operand, shape=(1 + context[0] + context[1], feature_dim))
+                    - m,
+                    s,
+                ),
+                shape=operand.shape,
+            )
+
+        return _func
+
+    def MyDNNLayer(hidden_size=128, num_layers=2):
+        return C.layers.Sequential(
+            [
+                C.layers.For(
+                    range(num_layers),
+                    lambda: C.layers.Dense(hidden_size, activation=C.sigmoid),
+                )
+            ]
+        )
+
+    def MyBLSTMLayer(hidden_size=128, num_layers=2):
+        W = C.Parameter(
+            (C.InferredDimension, hidden_size),
+            init=C.he_normal(1.0),
+            name="rnn_parameters",
+        )
+
+        def _func(operand):
+            return C.optimized_rnnstack(
+                operand,
+                weights=W,
+                hidden_size=hidden_size,
+                num_layers=num_layers,
+                bidirectional=True,
+                recurrent_op="lstm",
+            )
+
+        return _func
+
+    # Input variables denoting the features and label data
+    feature_var = C.sequence.input_variable(feature_dim * (1 + context[0] + context[1]))
+    label_var = C.sequence.input_variable(num_classes)
+
+    feature_norm = MyMeanVarNorm(feature_mean_file, feature_inv_stddev_file)(
+        feature_var
+    )
+    label_prior = load_ascii_vector(label_prior_file, "label_prior")
+    log_prior = C.log(label_prior)
+
     if model_type == "DNN":
-        model = DNNModel(globals["feature_dim"] * 23, 512, globals["num_classes"]).to(device)
+        net = MyDNNLayer(512, 4)(feature_norm)
+    elif model_type == "BLSTM":
+        net = MyBLSTMLayer(512, 2)(feature_norm)
     else:
-        model = BLSTMModel(globals["feature_dim"], 512, globals["num_classes"]).to(device)
+        raise RuntimeError("model_type must be DNN or BLSTM")
 
-    criterion = nn.CrossEntropyLoss(ignore_index=-100)
-    optimizer = optim.AdamW(model.parameters(), lr=1e-3, weight_decay=1e-5)
-    scheduler = optim.lr_scheduler.ReduceLROnPlateau(optimizer, mode='min', 
-                                                   factor=0.5, patience=3, verbose=True)
+    out = C.layers.Dense(num_classes, init=C.he_normal(scale=1 / 3))(net)
 
-    best_val_loss = float('inf')
-    val_losses = []
-    patience = 10
-    patience_counter = 0
+    # loss and metric
+    ce = C.cross_entropy_with_softmax(out, label_var)
+    pe = C.classification_error(out, label_var)
+    ScaledLogLikelihood = C.minus(out, log_prior, name="ScaledLogLikelihood")
 
-    for epoch in range(num_epochs):
-        # Training phase
-        model.train()
-        train_loss = 0
-        start_time = time.time()
-        
-        for batch_idx, (batch_features, batch_labels, lengths) in enumerate(train_loader):
-            batch_features = batch_features.to(device)
-            batch_labels = batch_labels.to(device)
+    # talk to the user
+    C.logging.log_number_of_parameters(out)
+    print()
 
-            optimizer.zero_grad()
-            
-            if model_type == "DNN":
-                batch_size, seq_len, feat_dim = batch_features.size()
-                batch_features = batch_features.reshape(-1, feat_dim)
-                batch_labels = batch_labels.reshape(-1)
-                outputs = model(batch_features)
-            else:
-                outputs = model(batch_features)
-                mask = torch.arange(batch_labels.size(1), device=device)[None, :] < lengths[:, None]
-                outputs = outputs[mask]
-                batch_labels = batch_labels[mask]
-            
-            loss = criterion(outputs, batch_labels)
-            loss.backward()
-            torch.nn.utils.clip_grad_norm_(model.parameters(), max_norm=5.0)
-            optimizer.step()
-            
-            train_loss += loss.item()
-            
-            if batch_idx % 50 == 0:
-                logging.info(f'Epoch: {epoch}, Batch: {batch_idx}, Loss: {loss.item():.4f}')
-
-        avg_train_loss = train_loss / len(train_loader)
-        epoch_time = time.time() - start_time
-        
-        # Validation phase
-        model.eval()
-        val_loss = 0
-        correct = 0
-        total = 0
-        
-        with torch.no_grad():
-            for batch_features, batch_labels, lengths in val_loader:
-                batch_features = batch_features.to(device)
-                batch_labels = batch_labels.to(device)
-                
-                if model_type == "DNN":
-                    batch_size, seq_len, feat_dim = batch_features.size()
-                    batch_features = batch_features.reshape(-1, feat_dim)
-                    batch_labels = batch_labels.reshape(-1)
-                    outputs = model(batch_features)
-                else:
-                    outputs = model(batch_features)
-                    mask = torch.arange(batch_labels.size(1), device=device)[None, :] < lengths[:, None]
-                    outputs = outputs[mask]
-                    batch_labels = batch_labels[mask]
-                
-                loss = criterion(outputs, batch_labels)
-                val_loss += loss.item()
-                
-                _, predicted = outputs.max(1)
-                total += batch_labels.size(0)
-                correct += predicted.eq(batch_labels).sum().item()
-        
-        avg_val_loss = val_loss / len(val_loader)
-        accuracy = 100. * correct / total
-        
-        logging.info(f'Epoch {epoch}: Train Loss = {avg_train_loss:.4f}, '
-                    f'Val Loss = {avg_val_loss:.4f}, Accuracy = {accuracy:.2f}%, '
-                    f'Time = {epoch_time:.2f}s')
-        
-        scheduler.step(avg_val_loss)
-        
-        # Early stopping
-        if avg_val_loss < best_val_loss:
-            best_val_loss = avg_val_loss
-            patience_counter = 0
-            # Save best model
-            save_checkpoint(model, optimizer, epoch, avg_val_loss, 
-                          os.path.join(am_path, f"{model_type}_CE.pt"), True)
-        else:
-            patience_counter += 1
-            if patience_counter >= patience:
-                logging.info(f'Early stopping triggered after {epoch + 1} epochs')
-                break
-    
-    return model
-
-def save_checkpoint(model: nn.Module, optimizer: optim.Optimizer, epoch: int, 
-                   loss: float, path: str, is_best: bool = False):
-    checkpoint = {
-        'epoch': epoch,
-        'model_state_dict': model.state_dict(),
-        'optimizer_state_dict': optimizer.state_dict(),
-        'loss': loss,
+    return {
+        "feature": feature_var,
+        "label": label_var,
+        "output": out,
+        "ScaledLogLikelihood": ScaledLogLikelihood,
+        "ce": ce,
+        "pe": pe,
+        "final_hidden": net,  # adding last hidden layer output for future use in CTC tutorial
     }
-    torch.save(checkpoint, path)
-    if is_best:
-        best_path = path.replace('.pt', '_best.pt')
-        torch.save(checkpoint, best_path)
+
+
+def create_trainer(network, progress_writers, epoch_size):
+    # Set learning parameters
+    lr_per_sample = [1.0e-4]  # transplanted schedule
+    mm_time_constant = [2500] * 200
+    lr_schedule = C.learning_rate_schedule(
+        lr_per_sample, unit=C.learners.UnitType.sample, epoch_size=epoch_size
+    )
+    mm_schedule = C.learners.momentum_as_time_constant_schedule(
+        mm_time_constant, epoch_size=epoch_size
+    )
+
+    momentum_sgd_learner = C.learners.momentum_sgd(
+        network["output"].parameters, lr_schedule, mm_schedule
+    )
+
+    # Create trainer
+    return C.Trainer(
+        network["output"],
+        (network["ce"], network["pe"]),
+        [momentum_sgd_learner],
+        progress_writers,
+    )
+
+
+def train_and_test(
+    network,
+    trainer,
+    train_source,
+    minibatch_size,
+    restore,
+    model_path,
+    model_name,
+    epoch_size,
+    cv_source,
+):
+    input_map = {
+        network["feature"]: train_source.streams.features,
+        network["label"]: train_source.streams.labels,
+    }
+    cv_input_map = (
+        None
+        if cv_source is None
+        else {
+            network["feature"]: cv_source.streams.features,
+            network["label"]: cv_source.streams.labels,
+        }
+    )
+    cv_checkpoint_interval = 5  # evaluate dev set after every N epochs
+
+    checkpoint_config = CheckpointConfig(
+        frequency=epoch_size,
+        filename=os.path.join(model_path, model_name),
+        restore=restore,
+        preserve_all=True,
+    )
+
+    cv_checkpoint_config = CrossValidationConfig(
+        cv_source,
+        model_inputs_to_streams=cv_input_map,
+        frequency=epoch_size * cv_checkpoint_interval,
+    )
+
+    # Train all minibatches
+    training_session(
+        trainer=trainer,
+        mb_source=train_source,
+        model_inputs_to_streams=input_map,
+        mb_size=minibatch_size_schedule(minibatch_size, epoch_size),
+        progress_frequency=epoch_size,
+        checkpoint_config=checkpoint_config,
+        cv_config=cv_checkpoint_config,
+    ).train()
+
+
+def train_network(
+    network,
+    features_file,
+    labels_file,
+    label_mapping_file,
+    max_epochs,
+    minibatch_size=[256],
+    restore=False,
+    log_to_file=None,
+    num_mbs_per_log=None,
+    gen_heartbeat=False,
+    profiling=False,
+    cv_features_file=None,
+    cv_labels_file=None,
+    epoch_size=None,
+    feature_dim=None,
+    num_classes=None,
+    model_path=None,
+    context=(0, 0),
+    frame_mode=False,
+    model_type=None,
+):
+
+    progress_writers = [
+        C.logging.ProgressPrinter(
+            freq=num_mbs_per_log,
+            tag="CE_Training",
+            log_to_file=log_to_file,
+            num_epochs=max_epochs,
+        )
+    ]
+
+    trainer = create_trainer(network, progress_writers, epoch_size)
+
+    train_source = create_mb_source(
+        features_file,
+        labels_file,
+        label_mapping_file,
+        feature_dim=feature_dim,
+        num_classes=num_classes,
+        max_sweeps=max_epochs,
+        context=context,
+        frame_mode=frame_mode,
+    )
+
+    cv_source = create_mb_source(
+        cv_features_file,
+        cv_labels_file,
+        label_mapping_file,
+        feature_dim=feature_dim,
+        num_classes=num_classes,
+        max_sweeps=1,
+        context=context,
+        frame_mode=frame_mode,
+    )
+
+    train_and_test(
+        network=network,
+        trainer=trainer,
+        train_source=train_source,
+        minibatch_size=minibatch_size,
+        restore=restore,
+        model_path=model_path,
+        model_name=model_type + "_CE",
+        epoch_size=epoch_size,
+        cv_source=cv_source,
+    )
+
 
 def main():
+
     parser = argparse.ArgumentParser()
-    parser.add_argument("-t", "--type", help="Network type to train (DNN or BLSTM)",
-                       default="DNN")
+    parser.add_argument(
+        "-t",
+        "--type",
+        help="Network type to train (DNN or BLSTM)",
+        required=True,
+        default="DNN",
+    )
     args = parser.parse_args()
-    
+
     model_type = str.upper(args.type)
-    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-    logging.info(f"Using device: {device}")
-    
-    # Load mean and variance normalization parameters
-    feature_mean = np.loadtxt(globals["feature_mean_file"]).astype(np.float32)
-    feature_invstd = np.loadtxt(globals["feature_invstddev_file"]).astype(np.float32)
-    
-    # Create datasets
-    context = (11,11) if model_type == "DNN" else (0,0)
-    train_dataset = SpeechDataset(globals["features_file"], globals["labels_file"],
-                                 feature_mean, feature_invstd, context)
-    val_dataset = SpeechDataset(globals["cv_features_file"], globals["cv_labels_file"],
-                               feature_mean, feature_invstd, context)
-    
-    logging.info(f"Loaded {len(train_dataset)} features and {globals['num_classes']} labels")
-    logging.info(f"Loaded {len(val_dataset)} features and {globals['num_classes']} labels")
-    
-    # Create data loaders
-    batch_size = 256 if model_type == "DNN" else 4096
-    train_loader = DataLoader(train_dataset, batch_size=batch_size, shuffle=True)
-    val_loader = DataLoader(val_dataset, batch_size=batch_size, shuffle=False)
-    
-    # Train model
-    max_epochs = 100 if model_type == "DNN" else 1
-    model = train_model(model_type, train_loader, val_loader, device, max_epochs)
-    
-    # Save final model
+
+    if model_type == "DNN":
+        mb_size = [256]
+        context_frames = (11, 11)
+        frame_mode = True
+        max_epochs = 100
+    elif model_type == "BLSTM":
+        mb_size = [4096]
+        context_frames = (0, 0)
+        frame_mode = False
+        max_epochs = 1
+    else:
+        raise RuntimeError("type must be DNN or BLSTM")
+
+    # mb_size = [256] if model_type == 'DNN' else [4096]
+    # context_frames = (11, 11) if model_type == "DNN" else (0, 0)
+
+    epoch_size = 1257284  # size of the corpus, 1 epoch = 1 pass thru the training data
+
+    if not os.path.exists(am_path):
+        raise RuntimeError("Cannot find am path, cannot proceed")
+
     model_path = os.path.join(am_path, model_type)
-    os.makedirs(model_path, exist_ok=True)
-    torch.save(model.state_dict(), os.path.join(model_path, f"{model_type}_CE.pt"))
+    if not os.path.exists(model_path):
+        os.makedirs(model_path)
+
+    log_file = os.path.join(model_path, "log")
+
+    feature_dim = globals["feature_dim"]
+
+    num_classes = globals["num_classes"]
+
+    feature_mean_file = globals["feature_mean_file"]
+    feature_inv_stddev_file = globals["feature_invstddev_file"]
+
+    label_mapping_file = globals["label_mapping_file"]
+    label_prior_file = globals["label_priors"]
+
+    features_file = globals["features_file"]
+    labels_file = globals["labels_file"]
+
+    cv_features_file = globals["cv_features_file"]
+    cv_labels_file = globals["cv_labels_file"]
+
+    C.debugging.set_computation_network_trace_level(1)
+    network = create_network(
+        feature_dim=feature_dim,
+        num_classes=num_classes,
+        feature_mean_file=feature_mean_file,
+        feature_inv_stddev_file=feature_inv_stddev_file,
+        label_prior_file=label_prior_file,
+        context=context_frames,
+        model_type=model_type,
+    )
+
+    train_network(
+        network,
+        features_file,
+        labels_file,
+        label_mapping_file,
+        cv_features_file=cv_features_file,
+        cv_labels_file=cv_labels_file,
+        max_epochs=max_epochs,
+        minibatch_size=mb_size,
+        restore=True,
+        log_to_file=log_file,
+        num_mbs_per_log=1000,
+        feature_dim=feature_dim,
+        num_classes=num_classes,
+        gen_heartbeat=False,
+        epoch_size=epoch_size,
+        model_path=model_path,
+        context=context_frames,
+        frame_mode=frame_mode,
+        model_type=model_type,
+    )
+    model = C.combine(network["ScaledLogLikelihood"], network["final_hidden"])
+    model.save(os.path.join(model_path, model_type + "_CE_forCTC"))
+
 
 if __name__ == "__main__":
     main()
